timingila: AutoStart change trigger infinite loop
Problem is that a change to EU.{i}.AutoRestart might trigger, under some
conditions, a loop where AutoStart keeps flapping (true/flase).
AutoStart is writable in timingila (SoftwareModules) and in Cthulhu.
A change of its value on one component DM have to trigger a change on the
other component to keep the values in sync.

To solve the problem, we need to prevent the circular updates (i.e when a
change is coming from one component, it cannot trigger a change as
a consequence of the other component updating the same config).

Closes PPW-467

Signed-off-by: Toon Caeyers <toon.caeyers_ext@softathome.com>